Multithreaded refactoring complete.  Still pretty messy and inefficient and some components aren't fully implemented (like RdrTransientHeap), but it is working.
Also fixed x64 builds.

+++ RenderLab/FreeList.h
@@ -55,7 +55,7 @@ public:
 
 	uint getId(T_object* pObj)
 	{
-		return pObj - m_objects;
+		return (uint)(pObj - m_objects);
 	}
 
 private:
